+++ pkg/controllers/apply_controller_test.go
@@ -1,3 +1,19 @@
+/*
+Copyright 2021 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
 package controllers
 
 import (
@@ -208,9 +224,11 @@ func TestApplyUnstructured(t *testing.T) {
 	testDeploymentDiffSpec := testDeployment.DeepCopy()
 	testDeploymentDiffSpec.Spec.MinReadySeconds = 0
 	rawDiffSpec, _ := json.Marshal(testDeploymentDiffSpec)
-	testManifestDiffSpec := workv1alpha1.Manifest{RawExtension: runtime.RawExtension{
-		Raw: rawDiffSpec,
-	}}
+	testManifestDiffSpec := workv1alpha1.Manifest{
+		RawExtension: runtime.RawExtension{
+			Raw: rawDiffSpec,
+		},
+	}
 	diffSpecObj, diffSpecDynamicClient, diffSpecHash := createObjAndDynamicClient(testManifestDiffSpec.Raw)
 
 	patchFailClient := fake.NewSimpleDynamicClient(runtime.NewScheme())
