+++ pkg/utils/common.go
@@ -1,3 +1,19 @@
+/*
+Copyright 2021 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
 package utils
 
 import (
@@ -5,8 +21,8 @@ import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
 )
 
-// UpsertOwnerRef create or inserts the owner reference to the object
-func UpsertOwnerRef(ref metav1.OwnerReference, object metav1.Object) {
+// AddOwnerRef creates or inserts the owner reference to the object
+func AddOwnerRef(ref metav1.OwnerReference, object metav1.Object) {
 	owners := object.GetOwnerReferences()
 	if idx := indexOwnerRef(owners, ref); idx == -1 {
 		owners = append(owners, ref)
